Fixing errors in samples folder
