+++ samples/samples.go
@@ -1,0 +1,38 @@
+// Copyright 2014-2018 Spectra Logic Corporation. All Rights Reserved.
+// Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+// this file except in compliance with the License. A copy of the License is located at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// or in the "license" file accompanying this file.
+// This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+// CONDITIONS OF ANY KIND, either express or implied. See the License for the
+// specific language governing permissions and limitations under the License.
+
+package main
+
+import "spectra/ds3_go_sdk/samples/functions"
+
+// Runs the various sample code. Each sample is self contained and can be run separately.
+func main() {
+
+    // Lists all existing buckets on the BP.
+    // Source code: getServiceSample.go
+    functions.GetServiceSample()
+
+    // Creates a bucket on the BP and puts sample files in the bucket.
+    // Source code: putBulkSample.go
+    functions.PutBulkSample()
+
+    // Gets a list of S3 objects in a bucket.
+    // Source code: getBucketSample.go
+    functions.GetBucketSample()
+
+    // Retrieve an object from a bucket.
+    // Source code: getObjectSample.go
+    functions.GetObjectSample()
+
+    // Retrieves several objects from a bucket on the BP.
+    // Source code: getBulkSample.go
+    functions.GetBulkSample()
+}
